refactor: turn forge runner test into integration tests
