+++ forge/src/multi_runner.rs
@@ -364,1078 +364,3 @@ impl MultiContractRunnerBuilder {
         self
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        decode::decode_console_logs,
-        test_helpers::{
-            filter::Filter, COMPILED, COMPILED_WITH_LIBS, EVM_OPTS, LIBS_PROJECT, PROJECT,
-            RE_PATH_SEPARATOR,
-        },
-    };
-    use foundry_config::{Config, RpcEndpoint, RpcEndpoints};
-    use foundry_evm::trace::TraceKind;
-    use std::env;
-
-    static TEST_OPTS: TestOptions = TestOptions {
-        fuzz_runs: 256,
-        fuzz_max_local_rejects: 1024,
-        fuzz_max_global_rejects: 65536,
-        fuzz_seed: None,
-        invariant_runs: 256,
-        invariant_depth: 15,
-        invariant_fail_on_revert: false,
-        invariant_call_override: false,
-    };
-
-    /// Builds a base runner
-    fn base_runner() -> MultiContractRunnerBuilder {
-        MultiContractRunnerBuilder::default().sender(EVM_OPTS.sender)
-    }
-
-    /// Builds a non-tracing runner
-    fn runner() -> MultiContractRunner {
-        let mut config = Config::with_root(PROJECT.root());
-        config.rpc_endpoints = rpc_endpoints();
-
-        base_runner()
-            .with_cheats_config(CheatsConfig::new(&config, &EVM_OPTS))
-            .build(
-                &PROJECT.paths.root,
-                (*COMPILED).clone(),
-                EVM_OPTS.evm_env_blocking(),
-                EVM_OPTS.clone(),
-            )
-            .unwrap()
-    }
-
-    /// Builds a tracing runner
-    fn tracing_runner() -> MultiContractRunner {
-        let mut opts = EVM_OPTS.clone();
-        opts.verbosity = 5;
-        base_runner()
-            .build(&PROJECT.paths.root, (*COMPILED).clone(), EVM_OPTS.evm_env_blocking(), opts)
-            .unwrap()
-    }
-
-    // Builds a runner that runs against forked state
-    fn forked_runner(rpc: &str) -> MultiContractRunner {
-        let mut opts = EVM_OPTS.clone();
-
-        opts.env.chain_id = None; // clear chain id so the correct one gets fetched from the RPC
-        opts.fork_url = Some(rpc.to_string());
-
-        let env = opts.evm_env_blocking();
-        let fork = opts.get_fork(&Default::default(), env.clone());
-
-        base_runner()
-            .with_fork(fork)
-            .build(&LIBS_PROJECT.paths.root, (*COMPILED_WITH_LIBS).clone(), env, opts)
-            .unwrap()
-    }
-
-    /// the RPC endpoints used during tests
-    fn rpc_endpoints() -> RpcEndpoints {
-        RpcEndpoints::new([
-            (
-                "rpcAlias",
-                RpcEndpoint::Url(
-                    "https://eth-mainnet.alchemyapi.io/v2/Lc7oIGYeL_QvInzI0Wiu_pOZZDEKBrdf"
-                        .to_string(),
-                ),
-            ),
-            ("rpcEnvAlias", RpcEndpoint::Env("${RPC_ENV_ALIAS}".to_string())),
-        ])
-    }
-
-    /// A helper to assert the outcome of multiple tests with helpful assert messages
-    fn assert_multiple(
-        actuals: &BTreeMap<String, SuiteResult>,
-        expecteds: BTreeMap<
-            &str,
-            Vec<(&str, bool, Option<String>, Option<Vec<String>>, Option<usize>)>,
-        >,
-    ) {
-        assert_eq!(
-            actuals.len(),
-            expecteds.len(),
-            "We did not run as many contracts as we expected"
-        );
-        for (contract_name, tests) in &expecteds {
-            assert!(
-                actuals.contains_key(*contract_name),
-                "We did not run the contract {}",
-                contract_name
-            );
-
-            assert_eq!(
-                actuals[*contract_name].len(),
-                expecteds[contract_name].len(),
-                "We did not run as many test functions as we expected for {}",
-                contract_name
-            );
-            for (test_name, should_pass, reason, expected_logs, expected_warning_count) in tests {
-                let logs =
-                    decode_console_logs(&actuals[*contract_name].test_results[*test_name].logs);
-
-                let warnings_count = &actuals[*contract_name].warnings.len();
-
-                if *should_pass {
-                    assert!(
-                        actuals[*contract_name].test_results[*test_name].success,
-                        "Test {} did not pass as expected.\nReason: {:?}\nLogs:\n{}",
-                        test_name,
-                        actuals[*contract_name].test_results[*test_name].reason,
-                        logs.join("\n")
-                    );
-                } else {
-                    assert!(
-                        !actuals[*contract_name].test_results[*test_name].success,
-                        "Test {} did not fail as expected.\nLogs:\n{}",
-                        test_name,
-                        logs.join("\n")
-                    );
-                    assert_eq!(
-                        actuals[*contract_name].test_results[*test_name].reason, *reason,
-                        "Failure reason for test {} did not match what we expected.",
-                        test_name
-                    );
-                }
-
-                if let Some(expected_logs) = expected_logs {
-                    assert!(
-                        logs.iter().eq(expected_logs.iter()),
-                        "Logs did not match for test {}.\nExpected:\n{}\n\nGot:\n{}",
-                        test_name,
-                        expected_logs.join("\n"),
-                        logs.join("\n")
-                    );
-                }
-
-                if let Some(expected_warning_count) = expected_warning_count {
-                    assert_eq!(
-                        warnings_count, expected_warning_count,
-                        "Test {} did not pass as expected. Expected:\n{}Got:\n{}",
-                        test_name, expected_warning_count, warnings_count
-                    );
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_core() {
-        let mut runner = runner();
-        let results = runner.test(&Filter::new(".*", ".*", ".*core"), None, TEST_OPTS).unwrap();
-
-        assert_multiple(
-            &results,
-            BTreeMap::from([
-                (
-                    "core/FailingSetup.t.sol:FailingSetupTest",
-                    vec![(
-                        "setUp()",
-                        false,
-                        Some("Setup failed: setup failed predictably".to_string()),
-                        None,
-                        None,
-                    )],
-                ),
-                (
-                    "core/MultipleSetup.t.sol:MultipleSetup",
-                    vec![(
-                        "setUp()",
-                        false,
-                        Some("Multiple setUp functions".to_string()),
-                        None,
-                        Some(1),
-                    )],
-                ),
-                (
-                    "core/Reverting.t.sol:RevertingTest",
-                    vec![("testFailRevert()", true, None, None, None)],
-                ),
-                (
-                    "core/SetupConsistency.t.sol:SetupConsistencyCheck",
-                    vec![
-                        ("testAdd()", true, None, None, None),
-                        ("testMultiply()", true, None, None, None),
-                    ],
-                ),
-                (
-                    "core/DSStyle.t.sol:DSStyleTest",
-                    vec![("testFailingAssertions()", true, None, None, None)],
-                ),
-                (
-                    "core/ContractEnvironment.t.sol:ContractEnvironmentTest",
-                    vec![
-                        ("testAddresses()", true, None, None, None),
-                        ("testEnvironment()", true, None, None, None),
-                    ],
-                ),
-                (
-                    "core/PaymentFailure.t.sol:PaymentFailureTest",
-                    vec![(
-                        "testCantPay()",
-                        false,
-                        Some("EvmError: Revert".to_string()),
-                        None,
-                        None,
-                    )],
-                ),
-                (
-                    "core/LibraryLinking.t.sol:LibraryLinkingTest",
-                    vec![
-                        ("testDirect()", true, None, None, None),
-                        ("testNested()", true, None, None, None),
-                    ],
-                ),
-                (
-                    "core/Abstract.t.sol:AbstractTest",
-                    vec![("testSomething()", true, None, None, None)],
-                ),
-            ]),
-        );
-    }
-
-    #[test]
-    fn test_logs() {
-        let mut runner = runner();
-        let results = runner.test(&Filter::new(".*", ".*", ".*logs"), None, TEST_OPTS).unwrap();
-
-        assert_multiple(
-            &results,
-            BTreeMap::from([
-                (
-                    "logs/DebugLogs.t.sol:DebugLogsTest",
-                    vec![
-                        (
-                            "test1()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "2".into()]),
-                            None,
-                        ),
-                        (
-                            "test2()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "3".into()]),
-                            None,
-                        ),
-                        (
-                            "testFailWithRequire()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "5".into()]),
-                            None,
-                        ),
-                        (
-                            "testFailWithRevert()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "4".into(), "100".into()]),
-                            None,
-                        ),
-                        (
-                            "testLog()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "Error: Assertion Failed".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogs()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "0x61626364".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogAddress()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "0x0000000000000000000000000000000000000001".into(),
-                            ]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes32()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "0x6162636400000000000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogInt()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "-31337".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "0x61626364".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogString()",
-                            true,
-                            None,
-                            Some(vec!["0".into(), "1".into(), "here".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogNamedAddress()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "address: 0x0000000000000000000000000000000000000001".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogNamedBytes32()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "abcd: 0x6162636400000000000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogNamedDecimalInt()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "amount: -0.000000000000031337".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogNamedDecimalUint()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "amount: 1.000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogNamedInt()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "amount: -31337".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogNamedUint()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "amount: 1000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogNamedBytes()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "abcd: 0x61626364".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogNamedString()",
-                            true,
-                            None,
-                            Some(vec![
-                                "0".into(),
-                                "1".into(),
-                                "key: val".into()]),
-                            None,
-                        ),
-                    ],
-                ),
-                (
-                    "logs/HardhatLogs.t.sol:HardhatLogsTest",
-                    vec![
-                        (
-                            "testInts()",
-                            true,
-                            None,
-                            Some(vec![
-                                "constructor".into(),
-                                "0".into(),
-                                "1".into(),
-                                "2".into(),
-                                "3".into(),
-                            ]),
-                            None,
-                        ),
-                        (
-                            "testMisc()",
-                            true,
-                            None,
-                            Some(vec![
-                                "constructor".into(),
-                                "testMisc 0x0000000000000000000000000000000000000001".into(),
-                                "testMisc 42".into(),
-                            ]),
-                            None,
-                        ),
-                        (
-                            "testStrings()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "testStrings".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLog()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "test".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogInt()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "-31337".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogUint()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "1".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogString()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "test".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBool()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "false".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogAddress()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x0000000000000000000000000000000000000001".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes1()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes2()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes3()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes4()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes5()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes6()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes7()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes8()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes9()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes10()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes11()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes12()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes13()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes14()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes15()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes16()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes17()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes18()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes19()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes20()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes21()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes22()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes23()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes24()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes25()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes26()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes27()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes28()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes29()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes30()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x610000000000000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes31()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x61000000000000000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testLogBytes32()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x6100000000000000000000000000000000000000000000000000000000000000".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogUint()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "1".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogString()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "test".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogBool()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "false".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogAddress()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "0x0000000000000000000000000000000000000001".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogFormatString()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "formatted log str=test".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogFormatUint()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "formatted log uint=1".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogFormatAddress()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "formatted log addr=0x0000000000000000000000000000000000000001".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogFormatMulti()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "formatted log str=test uint=1".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogFormatEscape()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "formatted log % test".into()]),
-                            None,
-                        ),
-                        (
-                            "testConsoleLogFormatSpill()",
-                            true,
-                            None,
-                            Some(vec!["constructor".into(), "formatted log test 1".into()]),
-                            None,
-                        ),
-                    ],
-                ),
-            ]),
-        );
-    }
-
-    #[test]
-    fn test_env_vars() {
-        let mut runner = runner();
-
-        // test `setEnv` first, and confirm that it can correctly set environment variables,
-        // so that we can use it in subsequent `env*` tests
-        runner.test(&Filter::new("testSetEnv", ".*", ".*"), None, TEST_OPTS).unwrap();
-        let env_var_key = "_foundryCheatcodeSetEnvTestKey";
-        let env_var_val = "_foundryCheatcodeSetEnvTestVal";
-        let res = env::var(env_var_key);
-        assert!(
-            res.is_ok() && res.unwrap() == env_var_val,
-            "Test `testSetEnv` did not pass as expected.
-Reason: `setEnv` failed to set an environment variable `{}={}`",
-            env_var_key,
-            env_var_val
-        );
-    }
-
-    /// Executes reverting fork test
-    #[test]
-    fn test_cheats_fork_revert() {
-        let mut runner = runner();
-        let suite_result = runner
-            .test(
-                &Filter::new(
-                    "testNonExistingContractRevert",
-                    ".*",
-                    &format!(".*cheats{}Fork", RE_PATH_SEPARATOR),
-                ),
-                None,
-                TEST_OPTS,
-            )
-            .unwrap();
-        assert_eq!(suite_result.len(), 1);
-
-        for (_, SuiteResult { test_results, .. }) in suite_result {
-            for (_, result) in test_results {
-                assert_eq!(
-                     result.reason.unwrap(),
-                     "Contract 0xCe71065D4017F316EC606Fe4422e11eB2c47c246 does not exists on active fork with id `1`\n        But exists on non active forks: `[0]`"
-                );
-            }
-        }
-    }
-
-    // <https://github.com/foundry-rs/foundry/issues/2623>
-    #[test]
-    fn test_issue_2623() {
-        let mut runner = runner();
-        let suite_result =
-            runner.test(&Filter::new(".*", ".*", ".*repros/Issue2623"), None, TEST_OPTS).unwrap();
-        assert!(!suite_result.is_empty());
-
-        for (_, SuiteResult { test_results, .. }) in suite_result {
-            for (test_name, result) in test_results {
-                let logs = decode_console_logs(&result.logs);
-                assert!(
-                    result.success,
-                    "Test {} did not pass as expected.\nReason: {:?}\nLogs:\n{}",
-                    test_name,
-                    result.reason,
-                    logs.join("\n")
-                );
-            }
-        }
-    }
-
-    /// Executes all non-reverting fork cheatcodes
-    #[test]
-    fn test_cheats_fork() {
-        let mut runner = runner();
-        let suite_result = runner
-            .test(
-                &Filter::new(".*", ".*", &format!(".*cheats{}Fork", RE_PATH_SEPARATOR))
-                    .exclude_tests(".*Revert"),
-                None,
-                TEST_OPTS,
-            )
-            .unwrap();
-        assert!(!suite_result.is_empty());
-
-        for (_, SuiteResult { test_results, .. }) in suite_result {
-            for (test_name, result) in test_results {
-                let logs = decode_console_logs(&result.logs);
-                assert!(
-                    result.success,
-                    "Test {} did not pass as expected.\nReason: {:?}\nLogs:\n{}",
-                    test_name,
-                    result.reason,
-                    logs.join("\n")
-                );
-            }
-        }
-    }
-
-    /// Executes all cheat code tests but not fork cheat codes
-    #[test]
-    fn test_cheats_local() {
-        let mut runner = runner();
-        let suite_result = runner
-            .test(
-                &Filter::new(".*", ".*", &format!(".*cheats{}[^Fork]", RE_PATH_SEPARATOR)),
-                None,
-                TEST_OPTS,
-            )
-            .unwrap();
-        assert!(!suite_result.is_empty());
-
-        for (_, SuiteResult { test_results, .. }) in suite_result {
-            for (test_name, result) in test_results {
-                let logs = decode_console_logs(&result.logs);
-                assert!(
-                    result.success,
-                    "Test {} did not pass as expected.\nReason: {:?}\nLogs:\n{}",
-                    test_name,
-                    result.reason,
-                    logs.join("\n")
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_fuzz() {
-        let mut runner = runner();
-
-        let suite_result =
-            runner.test(&Filter::new(".*", ".*", ".*fuzz/[^invariant]"), None, TEST_OPTS).unwrap();
-
-        assert!(!suite_result.is_empty());
-
-        for (_, SuiteResult { test_results, .. }) in suite_result {
-            for (test_name, result) in test_results {
-                let logs = decode_console_logs(&result.logs);
-
-                match test_name.as_str() {
-                    "testPositive(uint256)" |
-                    "testPositive(int256)" |
-                    "testSuccessfulFuzz(uint128,uint128)" |
-                    "testToStringFuzz(bytes32)" => assert!(
-                        result.success,
-                        "Test {} did not pass as expected.\nReason: {:?}\nLogs:\n{}",
-                        test_name,
-                        result.reason,
-                        logs.join("\n")
-                    ),
-                    _ => assert!(
-                        !result.success,
-                        "Test {} did not fail as expected.\nReason: {:?}\nLogs:\n{}",
-                        test_name,
-                        result.reason,
-                        logs.join("\n")
-                    ),
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_trace() {
-        let mut runner = tracing_runner();
-        let suite_result =
-            runner.test(&Filter::new(".*", ".*", ".*trace"), None, TEST_OPTS).unwrap();
-
-        // TODO: This trace test is very basic - it is probably a good candidate for snapshot
-        // testing.
-        for (_, SuiteResult { test_results, .. }) in suite_result {
-            for (test_name, result) in test_results {
-                let deployment_traces =
-                    result.traces.iter().filter(|(kind, _)| *kind == TraceKind::Deployment);
-                let setup_traces =
-                    result.traces.iter().filter(|(kind, _)| *kind == TraceKind::Setup);
-                let execution_traces =
-                    result.traces.iter().filter(|(kind, _)| *kind == TraceKind::Deployment);
-
-                assert_eq!(
-                    deployment_traces.count(),
-                    1,
-                    "Test {} did not have exactly 1 deployment trace.",
-                    test_name
-                );
-                assert!(
-                    setup_traces.count() <= 1,
-                    "Test {} had more than 1 setup trace.",
-                    test_name
-                );
-                assert_eq!(
-                    execution_traces.count(),
-                    1,
-                    "Test {} did not not have exactly 1 execution trace.",
-                    test_name
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_fork() {
-        let rpc_url = foundry_utils::rpc::next_http_archive_rpc_endpoint();
-        let mut runner = forked_runner(&rpc_url);
-        let suite_result =
-            runner.test(&Filter::new(".*", ".*", ".*fork"), None, TEST_OPTS).unwrap();
-
-        for (_, SuiteResult { test_results, .. }) in suite_result {
-            for (test_name, result) in test_results {
-                let logs = decode_console_logs(&result.logs);
-
-                assert!(
-                    result.success,
-                    "Test {} did not pass as expected.\nReason: {:?}\nLogs:\n{}",
-                    test_name,
-                    result.reason,
-                    logs.join("\n")
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_doesnt_run_abstract_contract() {
-        let mut runner = runner();
-        let results = runner
-            .test(
-                &Filter::new(".*", ".*", ".*Abstract.t.sol".to_string().as_str()),
-                None,
-                TEST_OPTS,
-            )
-            .unwrap();
-        assert!(results.get("core/Abstract.t.sol:AbstractTestBase").is_none());
-        assert!(results.get("core/Abstract.t.sol:AbstractTest").is_some());
-    }
-
-    #[test]
-    fn test_invariant() {
-        let mut runner = runner();
-
-        let mut opts = TEST_OPTS;
-        opts.invariant_call_override = true;
-        runner.test_options = opts;
-
-        let results =
-            runner.test(&Filter::new(".*", ".*", ".*fuzz/invariant/"), None, opts).unwrap();
-
-        assert_multiple(
-            &results,
-            BTreeMap::from([
-                (
-                    "fuzz/invariant/InvariantInnerContract.t.sol:InvariantInnerContract",
-                    vec![("invariantHideJesus", false, Some("jesus betrayed.".into()), None, None)],
-                ),
-                (
-                    "fuzz/invariant/InvariantReentrancy.t.sol:InvariantReentrancy",
-                    vec![("invariantNotStolen", false, Some("stolen.".into()), None, None)],
-                ),
-                (
-                    "fuzz/invariant/InvariantTest1.t.sol:InvariantTest",
-                    vec![("invariant_neverFalse", false, Some("false.".into()), None, None)],
-                ),
-                (
-                    "fuzz/invariant/target/ExcludeContracts.t.sol:ExcludeContracts",
-                    vec![("invariantTrueWorld", true, None, None, None)],
-                ),
-                (
-                    "fuzz/invariant/target/TargetContracts.t.sol:TargetContracts",
-                    vec![("invariantTrueWorld", true, None, None, None)],
-                ),
-                (
-                    "fuzz/invariant/target/TargetSenders.t.sol:TargetSenders",
-                    vec![("invariantTrueWorld", false, Some("false world.".into()), None, None)],
-                ),
-                (
-                    "fuzz/invariant/target/TargetSelectors.t.sol:TargetSelectors",
-                    vec![("invariantTrueWorld", true, None, None, None)],
-                ),
-            ]),
-        );
-    }
-}
