+++ forge/tests/it/main.rs
@@ -1,0 +1,10 @@
+mod cheats;
+pub mod config;
+mod core;
+mod fork;
+mod fuzz;
+mod invariant;
+mod repros;
+pub mod test_helpers;
+
+fn main() {}
